+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cleanup template formatting
 - Set `sensitive true` for `slapd.conf` template
 - Fix EL8 systemd unit for slapd daemon
+- Move platform attributes to library helpers
 
 ## 4.3.0 - *2020-11-23*